onboard/requirements.txt
altgraph==0.17.2
future==0.18.2
macholib==1.15.2
pika==1.3.1
pip==21.2.4
PyYAML==6.0
setuptools==58.0.4
six==1.15.0
wheel==0.37.0
